rearrange rails extensions, move JavaServletStore setup for Rails 3.x to a railtie initializer (closes issue #42)
