+++ src/main/ruby/jruby/rack/rails/extensions3.rb
@@ -5,9 +5,4 @@
 # See the file LICENSE.txt for details.
 #++
 
-require 'action_dispatch'
-module ActionDispatch
-  module Session
-    autoload :JavaServletStore, "action_dispatch/session/java_servlet_store"
-  end
-end
+require 'jruby/rack/rails/extensions'
